Split reprioritize snapshot tests
