+++ crates/oasis7/src/viewer/runtime_live/tests.rs
@@ -26,6 +26,7 @@ mod snapshot_generated_world;
 mod snapshot_micro_depot;
 mod snapshot_progress;
 mod snapshot_progress_empty_world;
+mod snapshot_reprioritize;
 #[path = "tests_support.rs"]
 mod tests_support;
